requirements.txt
# Core Framework
crewai>=0.80.0
crewai-tools>=0.2.6

# LLM Integration
langchain-google-genai>=2.0.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0

# Google AI
google-generativeai>=0.8.0

# OpenAI (for embeddings)
openai>=1.50.0

# Supabase Vector Database
supabase>=2.8.0
vecs>=0.4.5

# Web Interface
streamlit>=1.39.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0

# Visualizations
plotly>=5.20.0

# Machine Learning & Statistics
scikit-learn>=1.5.0
scipy>=1.14.0

# File Processing & Utilities
chardet>=5.2.0
pathlib2>=2.3.7
python-magic>=0.4.27

# Environment & Configuration
python-dotenv>=1.0.0
pydantic>=2.9.0
pydantic-settings>=2.5.0

# Utilities
requests>=2.32.0
urllib3>=2.2.0
aiohttp>=3.10.0

# Logging & Monitoring
loguru>=0.7.0

# Development & Testing (optional)
pytest>=8.3.0
pytest-asyncio>=0.24.0
black>=24.8.0
flake8>=7.1.0
mypy>=1.11.0

# Memory & Performance
psutil>=6.0.0
memory-profiler>=0.61.0

# Date & Time
python-dateutil>=2.9.0

# JSON & Data Serialization
ujson>=5.10.0

# HTTP & API
httpx>=0.27.0

# File Format Support
openpyxl>=3.1.5
xlrd>=2.0.1

# Cryptography (for secure operations)
cryptography>=43.0.0

# Additional Utilities
tqdm>=4.66.0
colorama>=0.4.6